Removed the unused function from transformUtils
